adapt method to code search

+++ src/ccaligner.py
@@ -2,19 +2,27 @@ import lexical_analysis.pretty_printing as prpr
 from launching import define_args
 from shutil import rmtree
 from clone_detection.algorithm import CCalignerAlgorithm
+import os
 
 args = define_args()
 
 codebase_loc = args.codebase_loc
 language = args.lang
+query_file = args.query_file
+
+if not os.path.isfile(query_file):
+    raise FileNotFoundError(f"Query file not found: {query_file}")
+
 if language == 'python':
     lang_ext = '.py'
 elif language == 'java':
     lang_ext = '.java'
-
+else:
+    raise ValueError(f"Unsupported language: {language}")
 
 pretty_loc = "./data/normalized_codebases/"
-rmtree(pretty_loc)
+if os.path.exists(pretty_loc):
+    rmtree(pretty_loc)
 
 
 pp = prpr.PrettyPrinter(codebase_loc, pretty_loc, language)
@@ -22,7 +30,7 @@ pp.pretty_print()
 
 final_dir = pretty_loc + codebase_loc.split('/')[-1] + '/' + 'obfuscated'
 
-cca = CCalignerAlgorithm(final_dir, lang_ext, 3, 0)
+cca = CCalignerAlgorithm(final_dir, lang_ext, 3, 0, query_file)
 pairs = cca.run_algo()
 for file1, file2 in pairs:
     file_name1 = file1.split('/')[-2]
